Move away from `Solo` namespace
