+++ osu.Game.Tests/Visual/Menus/TestSceneToolbarUserButton.cs
@@ -8,8 +8,8 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Testing;
+using osu.Game.Online;
 using osu.Game.Online.API;
-using osu.Game.Online.Solo;
 using osu.Game.Overlays.Toolbar;
 using osu.Game.Scoring;
 using osu.Game.Users;
