+++ osu.Game/Overlays/Toolbar/TransientUserStatisticsUpdateDisplay.cs
@@ -13,7 +13,7 @@ using osu.Framework.Threading;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
-using osu.Game.Online.Solo;
+using osu.Game.Online;
 using osu.Game.Resources.Localisation.Web;
 using osuTK;
