+++ osu.Game/Screens/Ranking/Statistics/User/OverallRanking.cs
@@ -6,7 +6,7 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Graphics.UserInterface;
-using osu.Game.Online.Solo;
+using osu.Game.Online;
 
 namespace osu.Game.Screens.Ranking.Statistics.User
 {
